requirements.txt
flasgger==0.9.5
Flask-RESTful==0.3.9
requests==2.28.1
pydantic==1.10.2
pymysql
flask-sqlalchemy==3.0.2